eAttestV3 cancellation fixes

+++ src/main/kotlin/org/taktik/freehealth/middleware/service/EattestV3Service.kt
@@ -69,7 +69,7 @@ interface EattestV3Service {
         patientFirstName: String,
         patientLastName: String,
         patientGender: String,
-        referenceDate: Int?,
+        referenceDate: Long?,
         eAttestRef: String,
         reason: String,
         attemptNbr: Int?
